% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-docs.R
\docType{package}
\name{codesamples-package}
\alias{codesamples}
\alias{codesamples-package}
\title{A large database of R snippets from Stack Overflow, Github and package
examples.}
\description{
\code{codesamples} contains three datasets:
\itemize{
\item \link{package_examples}
\item \link{so_questions}
\item \link{github_data}
}
}
\details{
All datasets have a \code{snippet} column which contains the actual R code.

checking it manually first!**
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hughjonesd/codesamples}
  \item Report bugs at \url{https://github.com/hughjonesd/codesamples/issues}
}

}
\author{
\strong{Maintainer}: David Hugh-Jones \email{davidhughjones@gmail.com}

}
\keyword{internal}
