% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-docs.R
\docType{data}
\name{github_data}
\alias{github_data}
\title{Github R Snippets}
\format{
\subsection{\code{github_data}}{

A data frame with 10,528 rows and 4 columns:
\describe{
\item{starter}{Starter for the github search}
\item{repo}{Github repo name}
\item{path}{Path within the repo}
\item{snippet}{The R snippet itself}
}
}
}
\usage{
github_data
}
\description{
R snippets from Github.
}
\details{
These were created by using the github search API
with \code{language:R} and a "starter"
string chosen randomly from \code{ls("package:base")}.

\strong{Obviously, don't run any code on your machine without
checking it manually first!}
}
\keyword{datasets}
