% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-docs.R
\docType{data}
\name{package_examples}
\alias{package_examples}
\title{Stack Overflow R Snippets}
\format{
\subsection{\code{package_examples}}{

A data frame with 1,544 rows and 3 columns:
\describe{
\item{package}{R package}
\item{topic}{Help topic}
\item{snippet}{The R code itself}
}
}
}
\usage{
package_examples
}
\description{
R examples from 100 randomly chosen packages.
}
\details{
Unlike the other two datasets, example code \emph{should} be
reasonably safe. Still it is best to \strong{not run any code on
your machine without checking it manually first!}
}
\keyword{datasets}
