% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-docs.R
\docType{data}
\name{so_questions}
\alias{so_questions}
\title{Stack Overflow R Snippets}
\format{
\subsection{\code{so_questions}}{

A data frame with 11,333 rows and 3 columns:
\describe{
\item{post_id}{SO post ID}
\item{creation_date}{Question creation date}
\item{snippet}{The R code itself}
}
}
}
\usage{
so_questions
}
\description{
R snippets from Stack Overflow questions.
}
\details{
Questions were from 2013 onwards, with a
minimum score of 3 and the \code{R} tag.

Note that not all code that parses is guaranteed
to be valid R. For example, a R \code{DESCRIPTION} file
may parse as R code.

\strong{Obviously, don't run any code on your machine without
checking it manually first!}

Here is the original SQL
query on <data.stackexchange.com>:

\if{html}{\out{<div class="sourceCode">}}\preformatted{SELECT DISTINCT
  p.Id,
  p.PostTypeId,
  p.Body,
  p.CreationDate
FROM
Posts p
INNER JOIN PostTags pt ON p.Id = pt.PostId
INNER JOIN Tags t ON pt.TagId = t.Id
WHERE
(UPPER(t.TagName) LIKE UPPER('R'))
AND
(p.Score >= 3)
AND
(p.CreationDate > '2013-01-01')
}\if{html}{\out{</div>}}
}
\keyword{datasets}
